+++ packages/handler/src/index.ts
@@ -4,22 +4,22 @@ export * from './BaseLogger'
 
 // ./commands
 
-export * from './commands/SlashCommandInterface'
+export * from './commands/SlashCommand'
 
-export * from './commands/UserCommandInterface'
+export * from './commands/UserCommand'
 
-export * from './commands/MessageCommandInterface'
+export * from './commands/MessageCommand'
 
 export * from './commands/CommandManager'
 
 // ./listeners
 
-export * from './listeners/ListenerInterface'
+export * from './listeners/Listener'
 
 export * from './listeners/ListenerManager'
 
 // ./tasks
 
-export * from './tasks/TaskInterface'
+export * from './tasks/Task'
 
 export * from './tasks/TaskManager'
